api/predict/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
numpy==2.1.2
pandas==2.2.2
pydantic==2.9.2
scikit-learn==1.4.2
prometheus_client==0.21.0
mlflow==2.18.0
python-dotenv==1.0.1
prometheus_fastapi_instrumentator==7.0.0
